Move room and client to separate files.
This moves the room and client classes to separate files/modules in
preparation for moving handlers to separate modules/files.
